feat: product polocy refactor: product page to implements policy, cashier page to flash message when error

+++ app/Services/ProductService.php
@@ -5,6 +5,7 @@ namespace App\Services;
 use App\Models\Product;
 use App\Repositories\ProductRepository;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\Gate;
 
 class ProductService
 {
@@ -15,23 +16,26 @@ class ProductService
         $this->productRepository = $productRepository;
     }
 
-    public function getAllProducts(string $category = null): Collection
+    public function getAllProducts(?string $category = null): Collection
     {
         return $this->productRepository->getAll($category);
     }
 
     public function createNewProduct(array $data): Product
     {
+        Gate::authorize('create', Product::class);
         return $this->productRepository->create($data);
     }
 
     public function updateProduct(Product $product, array $data): bool
     {
+        Gate::authorize('update', $product);
         return $this->productRepository->update($product, $data);
     }
 
     public function deleteProduct(Product $product): bool
     {
+        Gate::authorize('delete', $product);
         return $this->productRepository->delete($product);
     }
 }
